Calculation for fiber orientation
The local fiber orientation vector of each voxel in the example 3D CT voxel data matrix will be calculated through the 3D structure tensor method. The three components (X, Y and Z components) of the local fiber orientation vector of each fiber voxel will be saved in 3 Excel tables. The angle between the projection of each local fiber orientation vector on XY, XZ, and YZ plane and the X, X, and Y axis will be  saved in 3 Excel tables, too. The position of each vector component or angle of projection in the newly-saved 3D matrix is corresponding to the original position in the original 3D CT voxel matrix .